Add TF-TRT option for inference mode.
This should allow improved inference performance. TRT can be toggled on using --use_trt flag.
FP32, FP16, and INT8 options are available and can be selected with --precision. Depending on
the version of TRT being used, there may not be much speedup because many operations are not
yet supported. INT8 mode will most likely not work at all - we can disable this argument if we want.